Update bioSyntax_SETUP.sh
The current source-highlight version is 3.1.8_8 and the script will fail to locate it.

Hard coding software version in the script is not a good idea. Instead, we should use `/usr/local/opt/source-highlight/share/source-highlight/` to always redirect to the latest version.

The `less/bp_append.txt` path is also incorrect, and was fixed.

BTW: it seems zsh is not supported, please correct me if I was wrong.