Improve default comment-function behavior
Currently, any bioSyntax files within VSCode default to having the "Toggle Line Comment" function add a "//" at the beginning of the line, and the "Toggle Block Comment" encloses the selection in "/* ... */".  

This change disables block comments and changes the Toggle Line Comment to use "#" which is much more common in these types of files.